+++ src/Data.js
@@ -5,7 +5,7 @@ function getDefaultObject() {
         oddities: new Decimal(2),
         oddityGain: new Decimal(0),
         //derivatives
-        derivs: [{b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2)}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2)}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2)}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2)}],
+        derivs: [{b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2),u:true}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2),u:false}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2),u:false}, {b:new Decimal(0),amt:new Decimal(0),c:new Decimal(2),u:false}],
         //misc
         currentTab: 1,
         updateIDs: [0, 0, 0],