+++ changelog.md
@@ -6,4 +6,5 @@
 **v0.0.1** Created this repo, created the base files, added Saving, Loading, Import, and Export systems, and added a favicon. <br>
 **v0.0.1b** Created this Changelog. <br>
 **v0.0.2** Added the Dosis Font family, and the CSS code required to use the fonts within it on texts. <br>
-**v0.0.3** Added some simple UI and added the functionality for Derivative I <br>
+**v0.0.3** Added some simple UI and added the functionality for Derivative I. <br>
+**v0.0.4** Functionality for all Derivs, Derivative II can be unlocked, the unlock system, and the Deriv production system. <br>